# kahoot-robo-parts/docs/swagger.yaml
swagger: "2.0"
info:
  description: "This is a simple robo parts API."
  version: "1.0.0"
  title: "Kahoot Robo Parts"
  contact:
    email: "vic.blaga@gmail.com"
host: "localhost:4567"
...
schemes:
- "http"
definitions:
  Part:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      name:
        type: "string"
...
  AddPartRequest:
    type: "object"
    properties:
      name:
        type: "string"
  UpdatePartRequest:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      name:
        type: "string"
...
  PartResponse:
    type: "object"
    properties:
      status:
        type: "string"
        enum: [SUCCESS, ERROR]
      data:
        $ref: "#/definitions/Part"
...
  PartsResponse:
    type: "object"
    properties:
      status:
        type: "string"
        enum: ["SUCCESS", "ERROR"]
      data:
        type: array
        items:
          $ref: "#/definitions/Part"
paths:
...
  /parts:
    get:
      tags:
      - "parts"
      summary: "Get all parts"
      operationId: "getAll"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/PartsResponse"
    post:
      tags:
      - "parts"
      summary: "Add new part"
      operationId: "addPart"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Part object to be added"
        required: true
        schema:
          $ref: "#/definitions/AddPartRequest"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/PartResponse"
    put:
      tags:
      - "parts"
      summary: "Update part"
      operationId: "updatePart"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Part object to be updated"
        required: true
        schema:
          $ref: "#/definitions/UpdatePartRequest"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/PartResponse"
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"