# kahoot-robo-parts/robo-parts-rest/src/main/java/com/kahoot/robo/parts/rest/responses/PartResponse.java
package com.kahoot.robo.parts.rest.responses;

import com.google.gson.annotations.SerializedName;
import lombok.Value;

@Value
public class PartResponse {
    @SerializedName("id") private int id;
    @SerializedName("name") private String name;
}